api-modelos/requirements.txt
python-dotenv
ipykernel
requests
openai